src/python/requirements.txt
pyparsing >= 2.0.2
ansicolors >= 1.0.2
pygments >= 1.6